Organize the GraphQL typedefs
